Segregate MergeTreeIndexTextPreprocessor into its own file
The code will increase a bit and would make the MergeTreeIndexText too large.
